# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Jali API
  version: 0.1.0
  description: |
    Contract between the Jali mobile app and the backend (Laravel today; any
    implementation that passes the contract tests can replace it — see
    `docs/api/README.md`).

    Scope of this version: identity, push tokens, driver profile, and the
    on-demand ride endpoints from `RIDE_HAILING_PLAN.md`. Operations marked
    `x-jali-status: planned` are agreed designs that are not implemented yet;
    everything else is implemented and checked by `backend/tests/Feature/ApiContractTest.php`.

    Conventions
    - All amounts are integers in RWF.
    - Coordinates are WGS84 decimal degrees.
    - Timestamps are ISO-8601 strings.
    - Errors: 401 `{error}`; 403 `{error}`; 409 `{message}`; 422 Laravel validation `{message, errors}`;
      429 `{message}` with `Retry-After` (seconds).
    - Rate limits (S21.7): every route 120/min per user (per IP when signed out),
      headers `X-RateLimit-Limit` / `X-RateLimit-Remaining`. Stricter: sign-in 10/min per IP
      and 20/h per email; OTP request 5/h per phone; OTP verify 10/h per phone;
      ride requests 6/min per rider.
servers:
  - url: https://api.jali.rw/api
    description: Production
  - url: https://jali.stoka.rw/backend/public/api
    description: Dev
  - url: http://127.0.0.1:4010
    description: Prism mock (`npm run api:mock` in mobile/)
security:
  - bearerAuth: []
...
paths:
...
  /rides/{id}:
    parameters:
      - $ref: "#/components/parameters/RideId"
    get:
      tags: [Rides]
      operationId: getRide
      x-jali-story: S4.1
      summary: Ride detail (rider or assigned driver only)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ride" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /admin/rides/{id}:
    parameters:
      - $ref: "#/components/parameters/RideId"
    get:
      tags: [Admin rides]
      operationId: getAdminRide
      x-jali-story: S10.2
      x-jali-permission: manage-rides
      summary: Ride with timeline, fare breakdown and ratings
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminRideDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Sanctum personal access token from `/auth/login` or `/auth/login/google`.
  parameters:
...
    RideId:
      name: id
      in: path
      required: true
      schema: { type: integer }
  responses:
...
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
              message: { type: string }
    Forbidden:
      description: Missing permission
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
    NotFound:
      description: Not found (or not yours)
      content:
        application/json:
          schema:
            type: object
            properties:
              message: { type: string }
...
  schemas:
...
    VehicleClass:
      type: string
      enum: [moto, car, comfort, van]
...
    AdminPerson:
      type: [object, "null"]
      required: [id, name, phone]
      properties:
        id: { type: integer }
        name: { type: [string, "null"] }
        phone: { type: [string, "null"] }
    AdminRideSummary:
      type: object
      required: [id, status, vehicle_class, rider, driver, pickup, dropoff, quoted_fare, final_fare, flagged, requested_at]
      properties:
        id: { type: integer }
        status: { $ref: "#/components/schemas/RideStatus" }
        vehicle_class: { $ref: "#/components/schemas/VehicleClass" }
        rider: { $ref: "#/components/schemas/AdminPerson" }
        driver: { $ref: "#/components/schemas/AdminPerson" }
        pickup: { $ref: "#/components/schemas/Place" }
        dropoff: { $ref: "#/components/schemas/Place" }
        quoted_fare: { type: integer }
        final_fare: { type: [integer, "null"] }
        flagged: { type: boolean }
        requested_at: { type: [string, "null"], format: date-time }
    AdminRideDetail:
      allOf:
        - $ref: "#/components/schemas/AdminRideSummary"
        - type: object
          required: [payment_method, cancel_reason, cancelled_by, pin_attempts, vehicle, fare, timeline, ratings]
          properties:
            payment_method: { type: [string, "null"], enum: [cash, momo, null] }
            cancel_reason: { type: [string, "null"] }
            cancelled_by: { type: [string, "null"], enum: [rider, driver, system, null] }
            pin_attempts: { type: integer }
            vehicle:
              type: [object, "null"]
              properties:
                plate: { type: string }
                model: { type: string }
                color: { type: [string, "null"] }
            fare:
              type: object
              required: [est_distance_km, est_minutes, pickup_km, rate, driver_fare, service_fee, quoted_fare, cancel_fee, commission_pct, commission, final_fare]
              properties:
                est_distance_km: { type: number }
                est_minutes: { type: integer }
                pickup_km: { type: number }
                rate: { type: object, description: Driver's prices locked at request time }
                driver_fare: { type: integer }
                service_fee: { type: integer }
                quoted_fare: { type: integer }
                cancel_fee: { type: integer }
                commission_pct: { type: number }
                commission: { type: [integer, "null"] }
                final_fare: { type: [integer, "null"] }
            timeline:
              type: array
              items:
                type: object
                required: [type, actor, payload, at]
                properties:
                  type: { type: string, description: "requested, accepted, arrived, pin_failed, started, completed, cancelled, expired, adjusted, …" }
                  actor: { type: [string, "null"], description: "Name, or \"system\"" }
                  payload: { type: [object, array, "null"] }
                  at: { type: [string, "null"], format: date-time }
            ratings:
              type: array
              items:
                type: object
                required: [from, stars, tags, comment]
                properties:
                  from: { type: string, enum: [rider, driver] }
                  stars: { type: integer, minimum: 1, maximum: 5 }
                  tags: { type: array, items: { type: string } }
                  comment: { type: [string, "null"] }
...
    Place:
      type: object
      required: [lat, lng]
      properties:
        lat: { type: number }
        lng: { type: number }
        address: { type: [string, "null"] }
...
    RideStatus:
      type: string
      enum: [requested, accepted, arrived, in_progress, completed, declined, expired, cancelled_by_rider, cancelled_by_driver]
    Ride:
      type: object
      description: One ride as seen by the requesting user (rider or driver)
      required: [id, role, status, mode, vehicle_class, pickup, dropoff, est_distance_km, est_minutes, quoted_fare, driver_fare,
                 service_fee, final_fare, driver_earnings, cancel_fee, cancel_reason, start_pin, payment_method, expires_at,
                 driver, rider, my_rating, requested_at, accepted_at, arrived_at, started_at, completed_at, cancelled_at]
      properties:
        id: { type: integer }
        role: { type: string, enum: [rider, driver], description: Who is looking }
        status: { $ref: "#/components/schemas/RideStatus" }
        mode: { type: string, enum: [pick, broadcast] }
        vehicle_class: { $ref: "#/components/schemas/VehicleClass" }
        pickup: { $ref: "#/components/schemas/Place" }
        dropoff: { $ref: "#/components/schemas/Place" }
        est_distance_km: { type: number }
        est_minutes: { type: integer }
        quoted_fare: { type: integer, description: "Locked total the rider pays (driver fare + Jali fee). Broadcast: lowest offer until a driver accepts, then the winner's price" }
        max_fare: { type: [integer, "null"], description: "Broadcast: the rider's maximum" }
        driver_fare: { type: integer }
        service_fee: { type: integer }
        final_fare: { type: [integer, "null"] }
        driver_earnings: { type: [integer, "null"], description: "Driver view only: driver fare minus commission" }
        cancel_fee: { type: integer }
        cancel_reason: { type: [string, "null"] }
        start_pin: { type: [string, "null"], description: "Rider view only, while the ride is active" }
        payment_method: { type: [string, "null"], enum: [cash, momo, null] }
        expires_at: { type: [string, "null"], format: date-time, description: "While requested: when the driver's answer times out" }
        driver:
          oneOf:
            - type: "null"
            - type: object
              required: [id, name, photo, rating, phone, location, vehicle]
              properties:
                id: { type: integer }
                name: { type: string }
                photo: { type: [string, "null"] }
                rating: { type: number }
                phone: { type: [string, "null"], description: "Rider view, once accepted" }
                momo:
                  description: "Rider view when paying by MoMo, once the trip started (S7.1)"
                  type: [object, "null"]
                  properties:
                    number: { type: string }
                    name: { type: [string, "null"] }
                location:
                  oneOf:
                    - type: "null"
                    - type: object
                      required: [lat, lng]
                      properties:
                        lat: { type: number }
                        lng: { type: number }
                        heading: { type: [number, "null"] }
                        at: { type: [string, "null"], format: date-time }
                vehicle:
                  oneOf:
                    - type: "null"
                    - type: object
                      required: [class, model, color, plate, photo]
                      properties:
                        class: { $ref: "#/components/schemas/VehicleClass" }
                        model: { type: string }
                        color: { type: [string, "null"] }
                        plate: { type: string }
                        photo: { type: [string, "null"] }
        rider:
          description: Driver view only
          oneOf:
            - type: "null"
            - type: object
              required: [name, rating, phone]
              properties:
                name: { type: string }
                rating: { type: [number, "null"] }
                phone: { type: [string, "null"] }
        my_rating: { type: [integer, "null"], description: Stars I gave for this ride }
        requested_at: { type: [string, "null"], format: date-time }
        accepted_at: { type: [string, "null"], format: date-time }
        arrived_at: { type: [string, "null"], format: date-time }
        started_at: { type: [string, "null"], format: date-time }
        completed_at: { type: [string, "null"], format: date-time }
        cancelled_at: { type: [string, "null"], format: date-time }
...